# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "PDF Transaction WebAPI",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/organization": {
            "get": {
                "tags": [
                    "Partners"
                ],
                "summary": "Get organizations",
                "description": "Get organizations",
                "operationId": "get_organizations",
                "responses": {
                    "200": {
                        "description": "Successfully"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "401": {
                        "description": "Not Authorized"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            },
            "post": {
                "tags": [
                    "Partners"
                ],
                "summary": "Create organization",
                "description": "Create organization",
                "operationId": "create_organization",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "text"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successfully"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "401": {
                        "description": "Not Authorized"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    },
                    "409": {
                        "description": "Conflict"
                    }
                }
            }
        },
        "/api/organization/{id}": {
            "get": {
                "tags": [
                    "Partners"
                ],
                "summary": "Get organization",
                "description": "Get organization",
                "operationId": "get_organization",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "organization id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "401": {
                        "description": "Not Authorized"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Partners"
                ],
                "summary": "Update organization",
                "description": "Update organization",
                "operationId": "update_organization",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "organization id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [],
                                "properties": {
                                    "name": {
                                        "type": "text"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successfully"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "401": {
                        "description": "Not Authorized"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Partners"
                ],
                "summary": "Delete organization",
                "description": "Delete organization",
                "operationId": "delete_organization",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "organization id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "401": {
                        "description": "Not Authorized"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
...
    }
}